backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1
reportlab>=4.0
python-docx>=0.8
pandas>=2.2
